requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
greenlet>=3.0.0
